Add new version check
