+++ cmd/device-agent/main.go
@@ -2,7 +2,9 @@ package main
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"net/http"
 	"os"
 	"os/signal"
 	"path/filepath"
@@ -27,7 +29,8 @@ import (
 )
 
 const (
-	healthCheckInterval = 20 * time.Second // how often to healthcheck gateways
+	healthCheckInterval  = 20 * time.Second // how often to healthcheck gateways
+	versionCheckInterval = 1 * time.Hour    // how often to check for a new version of naisdevice
 )
 
 func handleSignals(log *logrus.Entry, cancel context.CancelFunc) {
@@ -148,11 +151,17 @@ func run(ctx context.Context, log *logrus.Entry, cfg *config.Config, notifier no
 	das := deviceagent.NewServer(ctx, log.WithField("component", "device-agent-server"), cfg, rc, notifier, stateMachine.SendEvent)
 	pb.RegisterDeviceAgentServer(grpcServer, das)
 
+	newVersionChannel := make(chan bool, 1)
+	go versionChecker(ctx, newVersionChannel, notifier, log)
+
 	go func() {
 		// This routine forwards status updates from the state machine to the device agent server
+		newVersionAvailable := false
 		for ctx.Err() == nil {
 			select {
+			case newVersionAvailable = <-newVersionChannel:
 			case s := <-statusChannel:
+				s.NewVersionAvailable = newVersionAvailable
 				das.UpdateAgentStatus(s)
 			case <-ctx.Done():
 				break
@@ -176,6 +185,30 @@ func run(ctx context.Context, log *logrus.Entry, cfg *config.Config, notifier no
 	return nil
 }
 
+func versionChecker(ctx context.Context, newVersionChannel chan<- bool, notifier notify.Notifier, logger logrus.FieldLogger) {
+	versionCheckTimer := time.NewTimer(versionCheckInterval)
+	for ctx.Err() == nil {
+		select {
+		case <-ctx.Done():
+			return
+		case <-versionCheckTimer.C:
+			newVersionAvailable, err := checkNewVersionAvailable(ctx)
+			if err != nil {
+				logger.Infof("check for new version: %s", err)
+				break
+			}
+
+			newVersionChannel <- newVersionAvailable
+			if newVersionAvailable {
+				notifier.Infof("New version of device agent available: https://doc.nais.io/device/update/")
+				versionCheckTimer.Stop()
+			} else {
+				versionCheckTimer.Reset(versionCheckInterval)
+			}
+		}
+	}
+}
+
 func helperHealthCheck(ctx context.Context, client pb.DeviceHelperClient) error {
 	helperHealthCheckCtx, helperHealthCheckCancel := context.WithTimeout(ctx, 5*time.Second)
 	defer helperHealthCheckCancel()
@@ -184,4 +217,34 @@ func helperHealthCheck(ctx context.Context, client pb.DeviceHelperClient) error 
 		return err
 	}
 	return nil
+}
+
+func checkNewVersionAvailable(ctx context.Context) (bool, error) {
+	type response struct {
+		Tag string `json:"tag_name"`
+	}
+
+	req, err := http.NewRequestWithContext(ctx, "GET", "https://api.github.com/repos/nais/device/releases/latest", nil)
+	if err != nil {
+		return false, err
+	}
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return false, fmt.Errorf("retrieve current release version: %s", err)
+	}
+
+	defer resp.Body.Close()
+
+	res := &response{}
+	decoder := json.NewDecoder(resp.Body)
+	err = decoder.Decode(res)
+	if err != nil {
+		return false, fmt.Errorf("unmarshal response: %s", err)
+	}
+
+	if version.Version != res.Tag {
+		return true, nil
+	}
+
+	return false, nil
 }
